Start implementing path drawing.

+++ PdfXenon/Document/Processor/PdfMatrix.cs
@@ -5,12 +5,12 @@ namespace PdfXenon.Standard
 {
     public class PdfMatrix
     {
-        private double _m11;
-        private double _m12;
-        private double _m21;
-        private double _m22;
-        private double _offsetX;
-        private double _offsetY;
+        private float _m11;
+        private float _m12;
+        private float _m21;
+        private float _m22;
+        private float _offsetX;
+        private float _offsetY;
 
         public PdfMatrix()
         {
@@ -19,31 +19,34 @@ namespace PdfXenon.Standard
             _m22 = 1;
         }
 
-        public void Translate(double x, double y)
+        public void Translate(float x, float y)
         {
             // Only need to adjust the offsets
             _offsetX += x;
             _offsetY += y;
         }
 
-        public void Scale(double scaleX, double scaleY)
+        public void Scale(float scaleX, float scaleY)
         {
             // Only need to adjust the X, Y values
             Multiply(scaleX, 0, 0, scaleY, 0, 0);
         }
 
-        public void RotateAt(double angle, double centreX, double centreY)
+        public void RotateAt(float angle, float centreX, float centreY)
         {
             // We take degrees but need to convert to using radians
-            double radians = (angle % 360) * (Math.PI / 180);
-            double sin = Math.Sin(radians);
-            double cos = Math.Cos(radians);
-            Multiply(cos, sin, -sin, cos,  (centreX * (1.0 - cos)) + (centreY * sin), (centreY * (1.0 - cos)) - (centreX * sin));
+            float radians = (float)((angle % 360) * (Math.PI / 180));
+            float sin = (float)Math.Sin(radians);
+            float cos = (float)Math.Cos(radians);
+            Multiply(cos, sin, 
+                    -sin, cos,  
+                    (float)(centreX * (1.0 - cos)) + (centreY * sin), 
+                    (float)(centreY * (1.0 - cos)) - (centreX * sin));
         }
 
         public void Transform(PdfPoint pt)
         {
-            double tempX = (pt.X * _m11) + (pt.Y * _m21) + _offsetX;
+            float tempX = (float)((pt.X * _m11) + (pt.Y * _m21) + _offsetX);
             pt.Y = (pt.X * _m12) + (pt.Y * _m22) + _offsetY;
             pt.X = tempX;
         }
@@ -60,11 +63,11 @@ namespace PdfXenon.Standard
                 Transform(pt);
         }
 
-        private void Multiply(double m11, double m12, double m21, double m22, double offsetX, double offsetY)
+        private void Multiply(float m11, float m12, float m21, float m22, float offsetX, float offsetY)
         {
-            double tm11 = (_m11 * m11) + (_m12 * m21);
-            double tm12 = (_m11 * m12) + (_m12 * m22);
-            double tm21 = (_m21 * m11) + (_m22 * m21);
+            float tm11 = (_m11 * m11) + (_m12 * m21);
+            float tm12 = (_m11 * m12) + (_m12 * m22);
+            float tm21 = (_m21 * m11) + (_m22 * m21);
             _m22 = (_m21 * m12) + (_m22 * m22);
             _m21 = tm21;
             _m12 = tm12;
